added a loop to check if the string is valid
A loop is added where it goes through each character in the string and checks if the character is valid hexadecimal ASCII character or not. If it is not, then through a branch statement it completely informs system to end the whole program. When it goes through all of the characters, it now ends the loop.  #checking if the character is valid or not is still to be added